change to decimal for position

+++ TOIFeedServer/Managers/TagManager.cs
@@ -34,8 +34,8 @@ namespace TOIFeedServer.Managers
 
             if (!int.TryParse(form["radius"][0], out var radius) || 
                 radius < 1 ||
-                !double.TryParse(form["longitude"][0], out var longitude) ||
-                !double.TryParse(form["latitude"][0], out var latitude))
+                !decimal.TryParse(form["longitude"][0], out var longitude) ||
+                !decimal.TryParse(form["latitude"][0], out var latitude))
                 return null;
             
             return new TagModel
